+++ CHANGES.rst
@@ -4,7 +4,16 @@
 Bug Fixes
 ---------
 
--
+jump
+~~~~
+
+- Updated the code for both NIR Snowballs and MIRI Showers. The snowball
+  flagging will now extend the saturated core of snowballs. Also,
+  circles are no longer used for snowballs preventing the huge circles
+  of flagged pixels from a glancing CR.
+  Shower code is completely new and is now able to find extended
+  emission far below the single pixel SNR. It also allows detected
+  showers to flag groups after the detection. [#144]
 
 ramp_fitting
 ~~~~~~~~~~~~
@@ -87,8 +96,9 @@ General
 Changes to API
 --------------
 
-- Added support for Quantities in models required for the RomanCAL pipeline. [#124]     
-  
+- Added support for Quantities in models required for the RomanCAL
+  pipeline. [#124]
+
 ramp_fitting
 ~~~~~~~~~~~~